requirements.txt
scikit-learn>=1.0.2
pandas>=1.3.5
numpy>=1.21.0
joblib>=1.1.0
Flask>=2.0.0
pyserial>=3.5
